fix: 쿠폰 삭제버튼 위치 변경
