+++ apps/desserbee-web/src/app/[lang]/(user)/owner/dashboard/coupon/page.tsx
@@ -48,8 +48,6 @@ export default function CouponPage() {
     if (storeUuid) fetchCoupons();
   }, [storeUuid]);
 
-  console.log(coupons);
-
   return (
     <div className="min-h-screen bg-[#EBEBEB]">
       <DashBoardHeader title="쿠폰 등록하기" />
